Added specific errors

+++ LAMPAPI/DeleteContact.php
@@ -10,11 +10,22 @@
 	}
     else
     {
-        $sql = "delete from CONTACTS where USERID = '$userId' and ID = '$contactId'";
-        if($result = $conn->query($sql) != TRUE)
-        {
-            returnWithError($conn->error);
-        }
+		// Verify contact exists
+		$sql = "select * from CONTACTS where USERID = '$userId' and ID = '$contactId'";
+		$result = $conn->query($sql);
+		if($result->num_rows < 1)
+		{
+			returnWithError("Contact not found (UserID: $userId, ContactID: $contactId)");
+		}
+		else
+		{
+			// Delete contact
+			$sql = "delete from CONTACTS where USERID = '$userId' and ID = '$contactId'";
+			if($result = $conn->query($sql) != TRUE)
+			{
+				returnWithError($conn->error);
+			}
+		}
         $conn->close();
     }
     returnWithError("");
